# backend/docs/openapi-v1.yaml
openapi: 3.0.3
info:
  title: Retoucher API
  version: 1.0.0-phase1
  description: |2

    Retoucher backend API.

    This API manages the full photo retouching workflow:

    - JWT authentication
    - Order creation and submission
    - Image upload
    - Editor assignment
    - Editor work and delivery
    - Supervisor review and QC
    - Client review and revision requests
    - Settlement and payment workflow
    - Order comments, threaded replies, rich annotations
    - Comment resolve/unresolve workflow
    - Activity logs
    - Notifications
    - Dashboard and summary endpoints
paths:
...
  /api/payments/refunds/{id}/approve/:
    post:
      operationId: payments_refunds_approve_create
      description: POST /api/payments/refunds/{id}/approve/
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: یک مقداد عدد یکتا که این Refund را شناسایی میکند.
        required: true
      tags:
      - payments
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefundApproveRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/RefundApproveRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/RefundApproveRequest'
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefundApprove'
          description: ''
  /api/payments/refunds/{id}/reject/:
    post:
      operationId: payments_refunds_reject_create
      description: POST /api/payments/refunds/{id}/reject/
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: یک مقداد عدد یکتا که این Refund را شناسایی میکند.
        required: true
      tags:
      - payments
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefundRejectRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/RefundRejectRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/RefundRejectRequest'
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefundReject'
          description: ''
  /api/payments/refunds/{refund_pk}/evidence/:
    get:
      operationId: payments_refunds_evidence_list
      description: GET /api/payments/refunds/{refund_pk}/evidence/
      parameters:
      - name: ordering
        required: false
        in: query
        description: کدام فیلد باید هنگام مرتب‌سازی نتایج استفاده شود.
        schema:
          type: string
      - in: path
        name: refund_pk
        schema:
          type: integer
        required: true
      - name: search
        required: false
        in: query
        description: یک عبارت جستجو.
        schema:
          type: string
      tags:
      - payments
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RefundEvidence'
          description: ''
    post:
      operationId: payments_refunds_evidence_create
      description: POST /api/payments/refunds/{refund_pk}/evidence/
      parameters:
      - in: path
        name: refund_pk
        schema:
          type: integer
        required: true
      tags:
      - payments
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/RefundEvidenceRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/RefundEvidenceRequest'
        required: true
      security:
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefundEvidence'
          description: ''
...
components:
  schemas:
...
    RefundApprove:
      type: object
      properties:
        approved_amount:
          type: integer
        note:
          type: string
    RefundApproveRequest:
      type: object
      properties:
        approved_amount:
          type: integer
        note:
          type: string
...
    RefundEvidence:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        file:
          type: string
          format: uri
        description:
          type: string
          maxLength: 255
        uploaded_by:
          type: integer
          readOnly: true
          nullable: true
        uploaded_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - file
      - id
      - uploaded_at
      - uploaded_by
    RefundEvidenceRequest:
      type: object
      properties:
        file:
          type: string
          format: binary
        description:
          type: string
          maxLength: 255
      required:
      - file
    RefundReject:
      type: object
      properties:
        note:
          type: string
    RefundRejectRequest:
      type: object
      properties:
        note:
          type: string
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...